google-cloud-vision/requirements.txt
google-cloud-vision>=3.7.4
lxml==4.9.1  # XMLを処理するためのライブラリ
